requirements/base.txt
django==2.0.6
django-environ==0.4.4
djangorestframework==3.8.2
django-authtools==1.6.0
psycopg2==2.7.5
